Terminado el manejo de errores de todos los metodos CRUD

+++ src/main/java/com/appclientes/springboot/backend/apirest/controllers/ClienteRestController.java
@@ -22,6 +22,8 @@ import org.springframework.web.bind.annotation.RestController;
 import com.appclientes.springboot.backend.apirest.models.entity.Cliente;
 import com.appclientes.springboot.backend.apirest.models.services.IClienteService;
 
+import ch.qos.logback.classic.Logger;
+
 @CrossOrigin(origins = {"http://localhost:4200"})
 @RestController
 @RequestMapping("/api")
@@ -71,27 +73,64 @@ public class ClienteRestController {
 		}catch(DataAccessException e) {
 			response.put("mensaje", "Error al realizar el insert en la base de datos");
 			response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
 		}
-		
-		return new ResponseEntity<Cliente>(clienteNew, HttpStatus.CREATED);
+		response.put("mensaje","El cliente ha sido creado con exito!");
+		response.put("cliente", clienteNew);
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED);
 	}
 	
 	@PutMapping("/clientes/{id}")
-	@ResponseStatus(HttpStatus.CREATED)
-	public Cliente update(@RequestBody Cliente cliente, @PathVariable Long id) {
-		Cliente clienteActual = clienteService.findById(id);
+	public ResponseEntity<?> update(@RequestBody Cliente cliente, @PathVariable Long id) {
 		
+		Cliente clienteActual = clienteService.findById(id);
+		Cliente clienteUpdated = null;
+		
+		Map<String, Object> response = new HashMap<>();
+		
+		if(clienteActual == null) {
+			response.put("mensaje", "Error: no se pudo editar, el cliente ID: ".concat(id.toString().concat(" no existe en la base de datos!")));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.NOT_FOUND);
+		}
+		
+		try {
 		clienteActual.setApellido(cliente.getApellido());
 		clienteActual.setNombre(cliente.getNombre());
 		clienteActual.setEmail(cliente.getEmail());
+		clienteActual.setCreateAt(cliente.getCreateAt());
 		
-		return clienteService.save(clienteActual);
+		clienteUpdated = clienteService.save(clienteActual);
+		
+		}catch(DataAccessException e) {
+			response.put("mensaje", "Error al actualizar el cliente en la base de datos");
+			response.put("error", e.getMessage().concat(": ").concat(e.getMostSpecificCause().getMessage()));
+			return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);
+		}
+		
+		response.put("mensaje","El cliente ha sido actualizado con exito!");
+		response.put("cliente", clienteUpdated);
+		
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.CREATED);
 	}
 	
 	@DeleteMapping("/clientes/{id}")
-	@ResponseStatus(HttpStatus.NO_CONTENT)
-	public void delete(@PathVariable Long id) {
+	public ResponseEntity<?> delete(@PathVariable Long id) {
+		
+		Cliente clienteEliminar = clienteService.findById(id);
+		Map<String, Object> response = new HashMap<>();
+		
+		if(clienteEliminar == null) {
+				response.put("mensaje", "Error: El cliente que intento eliminar no existe en la base de datos");
+				return new ResponseEntity<Map<String, Object>>(response, HttpStatus.INTERNAL_SERVER_ERROR);					
+				
+		}
 		clienteService.delete(id);
+		response.put("mensaje","El cliente ha sido eliminado con exito!");
+		
+		
+		
+		return new ResponseEntity<Map<String, Object>>(response, HttpStatus.OK);
+		
 	}
 }
